Rename command.description to command.descriptor for future use
